Refactor codebase to use the concept of services for clusters.
