+++ cluster/cluster.go
@@ -1,0 +1,12 @@
+package cluster
+
+import "time"
+
+type Cluster struct {
+	ID         string
+	Creator    string
+	Owner      string
+	Timeout    time.Time
+	Nodes      []*Node
+	EntryPoint string
+}
