--- daemon/constants.go
@@ -1,9 +1,0 @@
-package daemon
-
-import (
-	"time"
-)
-
-var (
-	DEFAULT_CLUSTER_TIMEOUT = time.Date(2222, 1, 1, 0, 0, 0, 0, time.UTC)
-)
