+++ daemon/restserver.go
@@ -4,10 +4,15 @@ import (
 	"context"
 	"encoding/json"
 	"errors"
-	"fmt"
+	"github.com/couchbaselabs/cbdynclusterd/cluster"
+	"github.com/couchbaselabs/cbdynclusterd/dyncontext"
+	"github.com/couchbaselabs/cbdynclusterd/service"
+	"github.com/couchbaselabs/cbdynclusterd/service/docker"
+	"github.com/couchbaselabs/cbdynclusterd/store"
 	"log"
 	"net/http"
 	"net/url"
+	"strconv"
 	"strings"
 	"time"
 
@@ -43,7 +48,7 @@ type NodeJSON struct {
 	IPv6Address          string `json:"ipv6_address"`
 }
 
-func jsonifyNode(node *Node) NodeJSON {
+func jsonifyNode(node *cluster.Node) NodeJSON {
 	return NodeJSON{
 		ID:                   node.ContainerID,
 		ContainerName:        node.ContainerName,
@@ -55,8 +60,8 @@ func jsonifyNode(node *Node) NodeJSON {
 	}
 }
 
-func UnjsonifyNode(jsonNode *NodeJSON) *Node {
-	return &Node{
+func UnjsonifyNode(jsonNode *NodeJSON) *cluster.Node {
+	return &cluster.Node{
 		ContainerID:          jsonNode.ID,
 		ContainerName:        jsonNode.ContainerName,
 		State:                jsonNode.State,
@@ -85,7 +90,7 @@ type ClusterJSON struct {
 	EntryPoint string     `json:"entry"`
 }
 
-func jsonifyCluster(cluster *Cluster) ClusterJSON {
+func jsonifyCluster(cluster *cluster.Cluster) ClusterJSON {
 	jsonCluster := ClusterJSON{
 		ID:         cluster.ID,
 		Creator:    cluster.Creator,
@@ -100,41 +105,6 @@ func jsonifyCluster(cluster *Cluster) ClusterJSON {
 	}
 
 	return jsonCluster
-}
-
-func UnjsonifyDockerHost(dockerHost *DockerHostJSON) (string, error) {
-	if dockerHost == nil || dockerHost.Hostname == "" || dockerHost.Port == "" {
-		return "", errors.New("Docker hostname or port is empty")
-	}
-	return fmt.Sprintf("%s:%s", dockerHost.Hostname, dockerHost.Port), nil
-}
-
-func UnjsonifyVersion(version *VersionJSON) (string, error) {
-	if version == nil || version.Version == "" {
-		return "", errors.New("cbdynclusterd version is empty")
-	}
-	return version.Version, nil
-}
-
-func UnjsonifyCluster(jsonCluster *ClusterJSON) (*Cluster, error) {
-	cluster := &Cluster{}
-	cluster.ID = jsonCluster.ID
-	cluster.Creator = jsonCluster.Creator
-	cluster.Owner = jsonCluster.Owner
-	cluster.EntryPoint = jsonCluster.EntryPoint
-
-	clusterTimeout, err := time.Parse(time.RFC3339, jsonCluster.Timeout)
-	if err != nil {
-		return nil, err
-	}
-	cluster.Timeout = clusterTimeout
-
-	for _, jsonNode := range jsonCluster.Nodes {
-		node := UnjsonifyNode(&jsonNode)
-		cluster.Nodes = append(cluster.Nodes, node)
-	}
-
-	return cluster, nil
 }
 
 func getHttpContext(r *http.Request) (context.Context, error) {
@@ -153,7 +123,7 @@ func getHttpContext(r *http.Request) (context.Context, error) {
 		ignoreOwnership = true
 	}
 
-	return NewContext(r.Context(), user, ignoreOwnership), nil
+	return dyncontext.NewContext(r.Context(), user, ignoreOwnership), nil
 }
 
 func writeJSONError(w http.ResponseWriter, err error) {
@@ -189,20 +159,20 @@ func readJsonRequest(r *http.Request, data interface{}) error {
 	return jsonDec.Decode(data)
 }
 
-func HttpRoot(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpRoot(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte("This is the cbdyncluster daemon!\n"))
 }
 
 type GetClustersJSON []ClusterJSON
 
-func HttpGetClusters(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpGetClusters(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
 		return
 	}
 
-	clusters, err := getAllClusters(reqCtx)
+	clusters, err := d.getAllClusters(reqCtx)
 	if err != nil {
 		writeJSONError(w, err)
 		return
@@ -210,8 +180,8 @@ func HttpGetClusters(w http.ResponseWriter, r *http.Request) {
 
 	jsonClusters := make(GetClustersJSON, 0)
 
-	for _, cluster := range clusters {
-		jsonCluster := jsonifyCluster(cluster)
+	for _, c := range clusters {
+		jsonCluster := jsonifyCluster(c)
 		jsonClusters = append(jsonClusters, jsonCluster)
 	}
 
@@ -246,7 +216,7 @@ type NewClusterJSON struct {
 	ID string `json:"id"`
 }
 
-func HttpCreateCluster(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpCreateCluster(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -260,9 +230,7 @@ func HttpCreateCluster(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	clusterOpts := ClusterOptions{
-		Timeout: 1 * time.Hour,
-	}
+	timeout := 1 * time.Hour
 
 	if reqData.Timeout != "" {
 		clusterTimeout, err := time.ParseDuration(reqData.Timeout)
@@ -271,37 +239,42 @@ func HttpCreateCluster(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		clusterOpts.Timeout = clusterTimeout
+		timeout = clusterTimeout
 	}
 
-	//Get/refresh alias repo
-	if err := GetConfigRepo(); err != nil {
-		log.Printf("Get config failed: %v", err)
+	if timeout < 0 {
+		writeJSONError(w, errors.New("must specify a valid timeout for the cluster"))
+		return
+	}
+	if timeout > 2*7*24*time.Hour {
+		writeJSONError(w, errors.New("cannot allocate clusters for longer than 2 weeks"))
+		return
+	}
+	clusterOpts := docker.AllocateClusterOptions{
+		Deadline: time.Now().Add(timeout),
 	}
 
 	for _, node := range reqData.Nodes {
-		finalVersion, err := aliasServerVersion(node.ServerVersion)
-		if err != nil {
-			writeJSONError(w, err)
-			return
-		}
-		nodeVersion, err := parseServerVersion(finalVersion, node.UseCommunityEdition)
-		if err != nil {
-			writeJSONError(w, err)
-			return
-		}
-
-		nodeOpts := NodeOptions{
-			Name:          node.Name,
-			Platform:      node.Platform,
-			ServerVersion: finalVersion,
-			VersionInfo:   nodeVersion,
+		nodeOpts := docker.CreateNodeOptions{
+			Name:                node.Name,
+			Platform:            node.Platform,
+			ServerVersion:       node.ServerVersion,
+			UseCommunityEdition: node.UseCommunityEdition,
 		}
 		clusterOpts.Nodes = append(clusterOpts.Nodes, nodeOpts)
 	}
 
-	clusterID, err := allocateCluster(reqCtx, clusterOpts)
+	clusterID, err := d.dockerService.AllocateCluster(reqCtx, clusterOpts)
 	if err != nil {
+		writeJSONError(w, err)
+		return
+	}
+
+	meta := store.ClusterMeta{
+		Owner:   dyncontext.ContextUser(reqCtx),
+		Timeout: clusterOpts.Deadline,
+	}
+	if err := d.metaStore.CreateClusterMeta(clusterID, meta); err != nil {
 		writeJSONError(w, err)
 		return
 	}
@@ -314,7 +287,7 @@ func HttpCreateCluster(w http.ResponseWriter, r *http.Request) {
 
 type GetClusterJSON ClusterJSON
 
-func HttpGetCluster(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpGetCluster(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -323,13 +296,13 @@ func HttpGetCluster(w http.ResponseWriter, r *http.Request) {
 
 	clusterID := mux.Vars(r)["cluster_id"]
 
-	cluster, err := getCluster(reqCtx, clusterID)
+	c, err := d.dockerService.GetCluster(reqCtx, clusterID)
 	if err != nil {
 		writeJSONError(w, err)
 		return
 	}
 
-	jsonCluster := jsonifyCluster(cluster)
+	jsonCluster := jsonifyCluster(c)
 
 	writeJsonResponse(w, jsonCluster)
 }
@@ -338,7 +311,7 @@ type UpdateClusterJSON struct {
 	Timeout string `json:"timeout"`
 }
 
-func HttpGetDockerHost(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpGetDockerHost(w http.ResponseWriter, r *http.Request) {
 	hostURI, err := url.Parse(dockerHost)
 	if err != nil {
 		writeJSONError(w, err)
@@ -366,7 +339,7 @@ func HttpGetVersion(w http.ResponseWriter, r *http.Request) {
 	return
 }
 
-func HttpSetupCluster(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpSetupCluster(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -382,33 +355,40 @@ func HttpSetupCluster(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	cluster, err := getCluster(reqCtx, clusterID)
+	c, err := d.dockerService.GetCluster(reqCtx, clusterID)
 	if err != nil {
 		writeJSONError(w, err)
 		return
 	}
-	if len(cluster.Nodes) != len(reqData.Services) {
+	if len(c.Nodes) != len(reqData.Services) {
 		writeJSONError(w, errors.New("services does not map to number of nodes"))
 		return
 	}
 
-	epnode, err := SetupCluster(&ClusterSetupOptions{
-		Nodes: cluster.Nodes,
-		Conf:  reqData,
+	epnode, err := d.dockerService.SetupCluster(&service.ClusterSetupOptions{
+		Nodes:               c.Nodes,
+		Services:            reqData.Services,
+		UseHostname:         reqData.UseHostname,
+		UseIpv6:             reqData.UseIpv6,
+		MemoryQuota:         strconv.Itoa(reqData.RamQuota),
+		User:                reqData.User,
+		StorageMode:         reqData.StorageMode,
+		Bucket:              reqData.Bucket,
+		UseDeveloperPreview: reqData.UseDeveloperPreview,
 	})
 	if err != nil {
 		writeJSONError(w, err)
 		return
 	}
 
-	cluster.EntryPoint = epnode
+	c.EntryPoint = epnode
 
-	jsonCluster := jsonifyCluster(cluster)
+	jsonCluster := jsonifyCluster(c)
 	writeJsonResponse(w, jsonCluster)
 	return
 }
 
-func HttpUpdateCluster(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpUpdateCluster(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -431,7 +411,7 @@ func HttpUpdateCluster(w http.ResponseWriter, r *http.Request) {
 			return
 		}
 
-		refreshCluster(reqCtx, clusterID, newTimeout)
+		d.refreshCluster(reqCtx, clusterID, newTimeout)
 
 		w.WriteHeader(200)
 		return
@@ -440,7 +420,7 @@ func HttpUpdateCluster(w http.ResponseWriter, r *http.Request) {
 	writeJSONError(w, errors.New("not sure what you wanted to do"))
 }
 
-func HttpDeleteCluster(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpDeleteCluster(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -449,7 +429,7 @@ func HttpDeleteCluster(w http.ResponseWriter, r *http.Request) {
 
 	clusterID := mux.Vars(r)["cluster_id"]
 
-	err = killCluster(reqCtx, clusterID)
+	err = d.dockerService.KillCluster(reqCtx, clusterID)
 	if err != nil {
 		writeJSONError(w, err)
 		return
@@ -459,16 +439,16 @@ func HttpDeleteCluster(w http.ResponseWriter, r *http.Request) {
 }
 
 type AddBucketJSON struct {
-	Name         string `json:"name"`
-	StorageMode  string `json:"storage_mode"`
-	RamQuota     int    `json:"ram_quota"`
-	UseHostname  bool   `json:"use_hostname"`
-	ReplicaCount int    `json:"replica_count"`
-	BucketType   string `json:"bucket_type"`
-	EvictionPolicy   string `json:"eviction_policy"`
+	Name           string `json:"name"`
+	StorageMode    string `json:"storage_mode"`
+	RamQuota       int    `json:"ram_quota"`
+	UseHostname    bool   `json:"use_hostname"`
+	ReplicaCount   int    `json:"replica_count"`
+	BucketType     string `json:"bucket_type"`
+	EvictionPolicy string `json:"eviction_policy"`
 }
 
-func HttpAddBucket(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpAddBucket(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -484,8 +464,14 @@ func HttpAddBucket(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	err = addBucket(reqCtx, clusterID, AddBucketOptions{
-		Conf: reqData,
+	err = d.dockerService.AddBucket(reqCtx, clusterID, service.AddBucketOptions{
+		Name:           reqData.Name,
+		StorageMode:    reqData.StorageMode,
+		RamQuota:       reqData.RamQuota,
+		UseHostname:    reqData.UseHostname,
+		ReplicaCount:   reqData.ReplicaCount,
+		BucketType:     reqData.BucketType,
+		EvictionPolicy: reqData.EvictionPolicy,
 	})
 	if err != nil {
 		writeJSONError(w, err)
@@ -500,7 +486,7 @@ type AddSampleBucketJSON struct {
 	UseHostname  bool   `json:"use_hostname"`
 }
 
-func HttpAddSampleBucket(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpAddSampleBucket(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -516,8 +502,9 @@ func HttpAddSampleBucket(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	err = addSampleBucket(reqCtx, clusterID, AddSampleOptions{
-		Conf: reqData,
+	err = d.dockerService.AddSampleBucket(reqCtx, clusterID, service.AddSampleOptions{
+		SampleBucket: reqData.SampleBucket,
+		UseHostname:  reqData.UseHostname,
 	})
 	if err != nil {
 		writeJSONError(w, err)
@@ -534,7 +521,7 @@ type AddCollectionJSON struct {
 	UseHostname bool   `json:"use_hostname"`
 }
 
-func HttpAddCollection(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpAddCollection(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -550,8 +537,11 @@ func HttpAddCollection(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	err = addCollection(reqCtx, clusterID, AddCollectionOptions{
-		Conf: reqData,
+	err = d.dockerService.AddCollection(reqCtx, clusterID, service.AddCollectionOptions{
+		Name:        reqData.Name,
+		ScopeName:   reqData.ScopeName,
+		BucketName:  reqData.BucketName,
+		UseHostname: reqData.UseHostname,
 	})
 	if err != nil {
 		writeJSONError(w, err)
@@ -572,7 +562,7 @@ type CertAuthResultJSON struct {
 	ClientCert []byte `json:"client_cert"`
 }
 
-func HttpSetupClientCertAuth(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpSetupClientCertAuth(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -588,15 +578,16 @@ func HttpSetupClientCertAuth(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	cluster, err := getCluster(reqCtx, clusterID)
+	c, err := d.dockerService.GetCluster(reqCtx, clusterID)
 	if err != nil {
 		writeJSONError(w, err)
 		return
 	}
 
-	certData, err := SetupCertAuth(SetupClientCertAuthOptions{
-		Nodes: cluster.Nodes,
-		Conf:  reqData,
+	certData, err := d.dockerService.SetupCertAuth(service.SetupClientCertAuthOptions{
+		Nodes:     c.Nodes,
+		UserName:  reqData.UserName,
+		UserEmail: reqData.UserEmail,
 	})
 	if err != nil {
 		writeJSONError(w, err)
@@ -620,7 +611,7 @@ type BuildImageResponseJSON struct {
 	ImageName string `json:"image_name"`
 }
 
-func HttpBuildImage(w http.ResponseWriter, r *http.Request) {
+func (d *daemon) HttpBuildImage(w http.ResponseWriter, r *http.Request) {
 	reqCtx, err := getHttpContext(r)
 	if err != nil {
 		writeJSONError(w, err)
@@ -634,39 +625,33 @@ func HttpBuildImage(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	nodeVersion, err := parseServerVersion(reqData.ServerVersion, reqData.UseCommunityEdition)
-	if err != nil {
-		writeJSONError(w, err)
-		return
-	}
-
-	err = ensureImageExists(reqCtx, nodeVersion, "")
+	image, err := d.dockerService.EnsureImageExists(reqCtx, reqData.ServerVersion, reqData.UseCommunityEdition)
 	if err != nil {
 		writeJSONError(w, err)
 		return
 	}
 
 	imageJSON := BuildImageResponseJSON{
-		ImageName: nodeVersion.toImageName(),
+		ImageName: image,
 	}
 	writeJsonResponse(w, imageJSON)
 }
 
-func createRESTRouter() *mux.Router {
+func (d *daemon) createRESTRouter() *mux.Router {
 	r := mux.NewRouter()
-	r.HandleFunc("/", HttpRoot)
-	r.HandleFunc("/docker-host", HttpGetDockerHost).Methods("GET")
+	r.HandleFunc("/", d.HttpRoot)
+	r.HandleFunc("/docker-host", d.HttpGetDockerHost).Methods("GET")
 	r.HandleFunc("/version", HttpGetVersion).Methods("GET")
-	r.HandleFunc("/clusters", HttpGetClusters).Methods("GET")
-	r.HandleFunc("/clusters", HttpCreateCluster).Methods("POST")
-	r.HandleFunc("/cluster/{cluster_id}", HttpGetCluster).Methods("GET")
-	r.HandleFunc("/cluster/{cluster_id}", HttpUpdateCluster).Methods("PUT")
-	r.HandleFunc("/cluster/{cluster_id}/setup", HttpSetupCluster).Methods("POST")
-	r.HandleFunc("/cluster/{cluster_id}", HttpDeleteCluster).Methods("DELETE")
-	r.HandleFunc("/cluster/{cluster_id}/add-bucket", HttpAddBucket).Methods("POST")
-	r.HandleFunc("/cluster/{cluster_id}/add-sample-bucket", HttpAddSampleBucket).Methods("POST")
-	r.HandleFunc("/cluster/{cluster_id}/add-collection", HttpAddCollection).Methods("POST")
-	r.HandleFunc("/cluster/{cluster_id}/setup-cert-auth", HttpSetupClientCertAuth).Methods("POST")
-	r.HandleFunc("/images", HttpBuildImage).Methods("POST")
+	r.HandleFunc("/clusters", d.HttpGetClusters).Methods("GET")
+	r.HandleFunc("/clusters", d.HttpCreateCluster).Methods("POST")
+	r.HandleFunc("/cluster/{cluster_id}", d.HttpGetCluster).Methods("GET")
+	r.HandleFunc("/cluster/{cluster_id}", d.HttpUpdateCluster).Methods("PUT")
+	r.HandleFunc("/cluster/{cluster_id}/setup", d.HttpSetupCluster).Methods("POST")
+	r.HandleFunc("/cluster/{cluster_id}", d.HttpDeleteCluster).Methods("DELETE")
+	r.HandleFunc("/cluster/{cluster_id}/add-bucket", d.HttpAddBucket).Methods("POST")
+	r.HandleFunc("/cluster/{cluster_id}/add-sample-bucket", d.HttpAddSampleBucket).Methods("POST")
+	r.HandleFunc("/cluster/{cluster_id}/add-collection", d.HttpAddCollection).Methods("POST")
+	r.HandleFunc("/cluster/{cluster_id}/setup-cert-auth", d.HttpSetupClientCertAuth).Methods("POST")
+	r.HandleFunc("/images", d.HttpBuildImage).Methods("POST")
 	return r
 }
