+++ helper/helper.go
@@ -3,35 +3,24 @@ package helper
 import (
 	"encoding/base64"
 	"errors"
-	"flag"
 	"fmt"
+	"github.com/google/uuid"
 	"io/ioutil"
 	"net/http"
-	"os"
 	"regexp"
 	"strconv"
 	"strings"
 	"time"
 
-	"github.com/couchbase/gocb"
-	"github.com/couchbase/gocb/cbft"
 	"github.com/golang/glog"
 )
 
 var (
-	IniFile                    *string
 	Slowlog                    bool
-	RestRetry                  = 60
-	RestTimeout                = 60 * time.Second
-	WaitTimeout                = 30 * time.Second
-	restInterval               = 3 * time.Second
 	PattActual, _              = regexp.Compile("\\s*vb_active_itm_memory:\\s*([0-9]+)")
 	PattActualReplica, _       = regexp.Compile("\\s*vb_replica_itm_memory:\\s*([0-9]+)")
 	PattUncompressed, _        = regexp.Compile("\\s*vb_active_itm_memory_uncompressed:\\s*([0-9]+)")
 	PattUncompressedReplica, _ = regexp.Compile("\\s*vb_replica_itm_memory_uncompressed:\\s*([0-9]+)")
-	StopWorkload               = false
-	WorkloadRunTime            = 10 * time.Second
-	WorkloadRampTime           = 10 * time.Second
 	SampleBucketsCount         = map[string]float64{
 		"travel-sample":  31591,
 		"beer-sample":    7303,
@@ -40,6 +29,10 @@ var (
 )
 
 const (
+	RestRetry          = 60
+	RestTimeout        = 60 * time.Second
+	WaitTimeout        = 30 * time.Second
+	restInterval       = 3 * time.Second
 	SshPort            = 22
 	RestPort           = 8091
 	N1qlPort           = 8093
@@ -136,169 +129,12 @@ type RestCall struct {
 	ContentType  string
 }
 
-func Usage() {
-	fmt.Fprintf(os.Stderr, "Test all: go test ./... -stderrthreshold=[INFO|WARN|FATAL] -log_dir=[string]\n"+
-		"Test a file(traceability_test.go): go test ./traceability/traceability_test.go  -stderrthreshold=[INFO|WARN|FATAL] -log_dir=[string]\n"+
-		"Test single(TestTraceability): go test -run TestTraceability  -stderrthreshold=[INFO|WARN|FATAL] -log_dir=[string]\n")
-	flag.PrintDefaults()
-	os.Exit(2)
-}
-
-func LoadData(bucketName, ep, username, password string, numItems int) (*gocb.Bucket, error) {
-	goCluster, err := gocb.Connect("couchbase://" + ep)
-	glog.Infof("connect to %s, %s with %s %s", ep, bucketName, username, password)
-	if err != nil {
-		return nil, err
-	}
-	goCluster.Authenticate(gocb.PasswordAuthenticator{
-		Username: username,
-		Password: password,
-	})
-	bucket, err := goCluster.OpenBucket(bucketName, "")
-	if err != nil {
-		return nil, err
-	}
-
-	for i := 0; i < numItems; i++ {
-		if err := Upsert(bucket, i); err != nil {
-			return nil, err
-		}
-	}
-	return bucket, nil
-}
-
-func Get(bucket *gocb.Bucket, i int) error {
-	kvData := KvData{}
-	_, err := bucket.Get(strconv.Itoa(i), &kvData)
-
-	if kvData.PropValue != "SampleValue"+strconv.Itoa(i) ||
-		kvData.SubFields.SubValue != "SampleSubvalue"+strconv.Itoa(i) ||
-		kvData.SubFields.RecurringField != "RecurringSubvalue" {
-		return errors.New("Not expected data from Get operation")
-	}
-	return err
-}
-
-func Upsert(bucket *gocb.Bucket, i int) error {
-	_, err := bucket.Upsert(strconv.Itoa(i),
-		KvData{
-			PropValue: "SampleValue" + strconv.Itoa(i),
-			SubFields: SubFields{
-				SubValue:       "SampleSubvalue" + strconv.Itoa(i),
-				RecurringField: "RecurringSubvalue",
-			},
-		}, 0)
-	return err
-}
-
-func Remove(bucket *gocb.Bucket, i int) error {
-	key := strconv.Itoa(i)
-	if _, err := bucket.Remove(key, 0); err != nil {
-		return err
-	}
-
-	var val string
-	if _, err := bucket.Get(key, &val); err != nil {
-		if err == gocb.ErrKeyNotFound {
-			return nil
-		}
-		return err
-	} else {
-		return errors.New("key " + key + " is expected to be removed but exists!")
-	}
-}
-
-func RemoveKeyIfExists(bucket *gocb.Bucket, key string) error {
-	var val interface{}
-	_, err := bucket.Get(key, &val)
-	if err != nil {
-		if err == gocb.ErrKeyNotFound {
-			return nil
-		} else {
-			return err
-		}
-	} else {
-		_, err := bucket.Remove(key, 0)
-		return err
-	}
-}
-
-func N1qlWorkload(bucket *gocb.Bucket, numItems int, chErr chan error) {
-	valIndex := 0
-	for !StopWorkload {
-		valIndex = (valIndex + 1) % numItems
-		statement := fmt.Sprintf("SELECT * from %s where PropValue = 'SampleValue%d'", bucket.Name(), valIndex)
-		query := gocb.NewN1qlQuery(statement)
-		rows, err := bucket.ExecuteN1qlQuery(query, nil)
-		if err != nil {
-			chErr <- err
-			return
-		}
-		if rows == nil {
-			chErr <- errors.New("Unexpected result from N1ql. rows is null")
-			return
-		}
-	}
-	chErr <- nil
-}
-
-func FtsWorkload(bucket *gocb.Bucket, ftsIndexName string, numItems int, chErr chan error) {
-	valIndex := 0
-	for !StopWorkload {
-		valIndex = (valIndex + 1) % numItems
-		str := fmt.Sprintf("SampleValue%d", valIndex)
-		query := gocb.NewSearchQuery(ftsIndexName, cbft.NewMatchQuery(fmt.Sprintf(str)))
-		res, err := bucket.ExecuteSearchQuery(query)
-		if err != nil {
-			chErr <- err
-			return
-		}
-
-		arrErr := res.Errors()
-
-		if len(arrErr) > 0 || len(res.Hits()) != 1 {
-			var msg string
-
-			if len(arrErr) > 0 {
-				msg += strings.Join(arrErr, ",")
-			}
-
-			if len(res.Hits()) != 1 {
-				msg += fmt.Sprintf("%s:Hits expected 1 but was %d", str, res.Hits())
-			}
-			chErr <- errors.New("Error from FTS:" + msg)
-			return
-		}
-	}
-	chErr <- nil
-}
-
-func KvGetWorkload(bucket *gocb.Bucket, numItems int, chErr chan error) {
-	valIndex := 0
-	for !StopWorkload {
-		valIndex = (valIndex + 1) % numItems
-		if err := Get(bucket, valIndex); err != nil {
-			chErr <- err
-			return
-		}
-	}
-	chErr <- nil
-}
-
-func KvUpsertWorkload(bucket *gocb.Bucket, numItems int, chErr chan error) {
-	valIndex := 0
-	for !StopWorkload {
-		valIndex = (valIndex + 1) % numItems
-		if err := Upsert(bucket, valIndex); err != nil {
-			chErr <- err
-			return
-		}
-	}
-	chErr <- nil
+func NewRandomClusterID() string {
+	uuid, _ := uuid.NewRandom()
+	return uuid.String()[0:8]
 }
 
 func RestRetryer(retry int, params *RestCall, fn func(*RestCall) (string, error)) (string, error) {
-
 	var body string
 	var err error
 	if body, err = fn(params); err != nil {
