+++ service/docker/certauth.go
@@ -1,4 +1,4 @@
-package daemon
+package docker
 
 import (
 	"crypto/rand"
@@ -6,46 +6,12 @@ import (
 	"crypto/x509"
 	"encoding/pem"
 	"fmt"
-	"strconv"
-	"time"
-
 	"github.com/couchbaselabs/cbcerthelper"
-
-	"github.com/couchbaselabs/cbdynclusterd/cluster"
-	"github.com/couchbaselabs/cbdynclusterd/helper"
+	"github.com/couchbaselabs/cbdynclusterd/service"
+	"time"
 )
 
-type SetupClientCertAuthOptions struct {
-	Nodes []*Node
-	Conf  SetupClientCertAuthJSON
-}
-
-type CertAuthResult struct {
-	CACert     []byte
-	ClientKey  []byte
-	ClientCert []byte
-}
-
-func SetupCertAuth(opts SetupClientCertAuthOptions) (*CertAuthResult, error) {
-	initialNodes := opts.Nodes
-	var nodes []cluster.Node
-	for i := 0; i < len(initialNodes); i++ {
-		ipv4 := initialNodes[i].IPv4Address
-		hostname := ipv4
-
-		nodeHost := cluster.Node{
-			HostName:  hostname,
-			Port:      strconv.Itoa(helper.RestPort),
-			SshLogin:  &helper.Cred{Username: helper.SshUser, Password: helper.SshPass, Hostname: ipv4, Port: helper.SshPort},
-			RestLogin: &helper.Cred{Username: helper.RestUser, Password: helper.RestPass, Hostname: ipv4, Port: helper.RestPort},
-		}
-		nodes = append(nodes, nodeHost)
-	}
-
-	return setupCertAuth(opts.Conf.UserName, opts.Conf.UserEmail, nodes)
-}
-
-func setupCertAuth(username, email string, nodes []cluster.Node) (*CertAuthResult, error) {
+func (ds *DockerService) setupCertAuth(username, email string, nodes []Node) (*service.CertAuthResult, error) {
 	rootKey, err := rsa.GenerateKey(rand.Reader, 2048)
 	if err != nil {
 		return nil, fmt.Errorf("failed to generate private key: %v", err)
@@ -84,7 +50,7 @@ func setupCertAuth(username, email string, nodes []cluster.Node) (*CertAuthResul
 	keyOut := pem.EncodeToMemory(&pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(clientKey)})
 	clientOut := pem.EncodeToMemory(&pem.Block{Type: "CERTIFICATE", Bytes: clientCertBytes})
 
-	return &CertAuthResult{
+	return &service.CertAuthResult{
 		CACert:     rootOut,
 		ClientKey:  keyOut,
 		ClientCert: clientOut,
