+++ service/docker/options.go
@@ -1,0 +1,10 @@
+package docker
+
+import (
+	"time"
+)
+
+type AllocateClusterOptions struct {
+	Deadline time.Time
+	Nodes    []CreateNodeOptions
+}
